+++ tests/test_models.py
@@ -1,12 +1,12 @@
 # coding: utf-8
+import django_tables2 as tables
 import pytest
 from django.db.models.functions import Length
 from django.utils import six
-
-import django_tables2 as tables
+from django_tables2 import RequestConfig
 
 from .app.models import Occupation, Person, PersonProxy
-from .utils import build_request
+from .utils import assertNumQueries, build_request
 
 pytestmark = pytest.mark.django_db
 request = build_request('/')
@@ -57,7 +57,7 @@ def test_model_table():
 
     class ComplexModel(models.Model):
         char = models.CharField(max_length=200)
-        fk = models.ForeignKey('self')
+        fk = models.ForeignKey('self', on_delete=models.CASCADE)
         m2m = models.ManyToManyField('self')
 
         class Meta:
@@ -119,6 +119,7 @@ def test_column_verbose_name():
     # however both fields that use the ``first_name`` field should just use a
     # titlised version of the column name as the column header.
     table = PersonTable(Person.objects.all())
+
     # Should be generated (capitalized column name)
     assert 'First Name' == table.columns['first_name'].verbose_name
     assert 'First Name' == table.columns['fn1'].verbose_name
@@ -147,7 +148,7 @@ def test_column_verbose_name():
             model = Person
 
     # Issue #16
-    table = PersonTable([])
+    table = PersonTable(Person.objects.all())
     assert 'Translation Test' == table.columns['trans_test'].verbose_name
     assert 'Translation Test Lazy' == table.columns['trans_test_lazy'].verbose_name
     assert 'Web Site' == table.columns['website'].verbose_name
@@ -401,3 +402,23 @@ def test_column_named_delete():
 
     assert Person.objects.get(pk=person1.pk) == person1
     assert Person.objects.get(pk=person2.pk) == person2
+
+
+def test_single_query_for_non_paginated_table():
+    '''
+    A non-paginated table should not generate a query for each row, but only
+    one query fetch the rows.
+    '''
+    for i in range(10):
+        Person.objects.create(first_name='Bob %d' % i, last_name='Builder')
+
+    class PersonTable(tables.Table):
+        class Meta:
+            model = Person
+            fields = ('first_name', 'last_name')
+            order_by = ('last_name', 'first_name')
+
+    table = PersonTable(Person.objects.all())
+
+    with assertNumQueries(1):
+        list(table.as_values())